+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.7.0] - 2021-07-08
+
+### Added
+
+- IKeyboardShortcutsService with method for retrieving Keyboard Shortucts for workspace. - [RTF-1257](https://jira.kcura.com/browse/RTF-1257)
+
+### Changed
+
+- RTF version changed to v4.5.0. - [RTF-1257](https://jira.kcura.com/browse/RTF-1257)
+
 ## [3.6.0] - 2021-07-07
 
 ### Changed